requirements.txt
aiofiles                  
aiohttp                   
aiosignal                 
altair                    
annotated-types           
anyio                     
asgiref                   
attrs                     
bardapi                   
bcrypt                    
blinker                   
browser-cookie3           
cachetools                
certifi                   
charset-normalizer        
click                     
colorama                  
contourpy                 
cycler                    
dataclasses-json          
distlib                   
distro                    
dnspython                 
email_validator           
fastapi                   
fastapi-cli               
ffmpy                     
filelock                  
Flask                     
Flask-Bcrypt              
Flask-Login               
Flask-SQLAlchemy          
Flask-WTF                 
fonttools                 
frozenlist                
fsspec                    
gitdb                     
GitPython                 
gradio                    
gradio_client             
greenlet                  
h11                       
h2                        
hpack                     
httpcore                  
httptools                 
httpx                     
huggingface-hub           
hyperframe                
idna                      
importlib_resources       
itsdangerous              
Jinja2                    
joblib                    
jsonpatch                 
jsonpointer               
jsonschema                
jsonschema-specifications 
kiwisolver                
langchain                 
langchain-community       
langchain-core            
langchain-text-splitters  
langsmith                 
loguru                    
lz4                       
markdown-it-py            
MarkupSafe                
marshmallow               
matplotlib                
mdurl                     
multidict                 
mypy-extensions           
nltk                      
numpy                     
ollama                    
openai                    
orjson                    
packaging                 
pandas                    
pillow                    
pip                       
pipenv                    
platformdirs              
protobuf                  
pyarrow                   
pycryptodomex             
pydantic                  
pydantic_core             
pydeck                    
pydub                     
Pygments                  
pyparsing                 
python-dateutil           
python-dotenv             
python-gemini-api         
python-multipart          
pytz                      
PyYAML                    
referencing               
regex                     
requests                  
rich                      
rpds-py                   
ruff                      
semantic-version          
setuptools                
shellingham               
six                       
smmap                     
sniffio                   
SQLAlchemy                
sqlparse                  
starlette                 
streamlit                 
tenacity                  
toml                      
tomlkit                   
toolz                     
tornado                   
tqdm                      
typer                     
typing_extensions         
typing-inspect            
tzdata                    
ujson                     
urllib3                   
uvicorn                   
virtualenv                
watchdog                  
watchfiles                
websockets                
Werkzeug                  
win32-setctime            
WTForms                   
yarl                      
